% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenDivide.R
\name{annualScreen}
\alias{annualScreen}
\title{Break Up Canvas}
\usage{
annualScreen()
}
\arguments{
\item{empty}{No arguments required.}
}
\description{
This function allows you to break up the screen canvas into 2 columns of 2 for funnel plots and a column of 4 rows for density plots.
}
\examples{
zero_one()
}
\keyword{scale}
