% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data.R
\name{arrange.data}
\alias{arrange.data}
\title{Simplify data and add climate index values.}
\usage{
arrange.data(df, response)
}
\arguments{
\item{df}{Requires a list with nested response variables with a further nested dataframe of model output. This dataframe must contain 7 initial columns of model results with all subsequent columns containing climate data.}

\item{response}{Requires a character object referencing one of the nested response variables in df.}
}
\description{
This function allows you to summarize climate data into a single index values across variables. It requires the zero_one() function.
}
\examples{
arrange.data(df = out_doy2, response = "Area")
}
\keyword{climate}
\keyword{index}
