% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.density.R
\name{axis.density.draw}
\alias{axis.density.draw}
\title{Density Plot Axis' and Labels.}
\usage{
axis.density.draw(type, ALS, LL, LW, annual = F)
}
\arguments{
\item{type}{Requires a character vector of "int" if plotting intercept data. Anything less (e.g., "exp") will default to plotting variance exponent data.}

\item{ALS}{Requires a numeric that defines the axis line size.}

\item{LL}{Requires a numeric that defines the margin line where the x-axis label will be drawn.}

\item{LW}{Requires a numeric that defines the line width.}

\item{annual}{Requires a T/F as to whether the data are annually summarized (T) or monthly summarized (F). Defaults to False.}
}
\description{
This function allows you to draw the density plot axis' and labels.
}
\examples{
axis.density(type = "exp", ALS = 1, LL = 1)
}
\keyword{axis,}
\keyword{labels}
