% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_Fun.R
\name{axis.Fun}
\alias{axis.Fun}
\title{Y-Axis Labels}
\usage{
axis.Fun(df, yaxis.DOY = F, ylimit = NULL)
}
\arguments{
\item{df}{Requires a list with a nested dataframe of model results.}

\item{yaxis.DOY}{Requires a boolean T/F. This argument refers to whether or not the y-axis is logit transformed DOY or not. It defaults to False.}

\item{ylimit}{Allows the user to override estimates and force the bounds of the y-axis.}
}
\description{
Manually or automatically define readable y-axis labels for logit transformed day of year (DOY) trend variables or log transformed trend variables.
}
\examples{
axis.Fun(df, yaxis.DOY = F, ylimit = c(-0.04,0.04))
}
\keyword{y-axis}
