% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_density_data.R
\name{gather.density.data}
\alias{gather.density.data}
\title{Gather and Arrange Density Plot Data}
\usage{
gather.density.data(df, response, type)
}
\arguments{
\item{df}{Requires a list with nested response variables with a further nested dataframe of model output.}

\item{response}{Requires a character object referencing one of the nested response variables in df.}

\item{type}{Requires a character vector of "int" if plotting intercept data. Anything less (e.g., "exp") will default to plotting variance exponent data.}
}
\description{
Gather and arrange observed and simulated variance and intercept parameter estimates.
}
\examples{
gather.density.data(df = out_doy2, response = "Area", type = "int")
}
\keyword{density}
