% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sims.R
\name{line.sims}
\alias{line.sims}
\title{Draw Simulation Lines}
\usage{
line.sims(df, col, LW, shade = "10", transp = T, intercept, slope,
  resp.sqrt, sigma)
}
\arguments{
\item{df}{A dataframe containing a column of variance exponent parameters.}

\item{col}{A character object defining the colour of the lines.}

\item{LW}{A numeric object defining the width of each variance line.}

\item{shade}{Requires a character object between 0 and 100. The lower the value the greater the transparency. Default is "10".}

\item{transp}{A on/off (T/F) switch for making the lines transparent.}

\item{intercept}{Requires a numeric object with the observed intercept.}

\item{slope}{Requires a numeric object with the observed slope.}

\item{resp.sqrt}{Requires a numeric object with the observed response variables. This variable (resp.sqrt) was added by arrange.data().}

\item{sigma}{Requires a numeric object with the observed sigma.}
}
\description{
This function allows you to add simulation lines to plotting region.
}
\examples{
line.sims(df, col = "#000000", LW = 1, shade = "10", transp = T, intercep = 1, slope = 1, resp.sqrt = df$resp.sqrt, sigma = 2)
}
\keyword{simlines}
