% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenDivide.R
\name{monthlyScreen}
\alias{monthlyScreen}
\title{Break Up Canvas}
\usage{
monthlyScreen()
}
\arguments{
\item{empty}{No arguments required.}
}
\description{
This function allows you to break up the screen canvas into two columns of twelve for density plots and a column of 4 rows for representitive seasonal funnel plots.
}
\examples{
zero_one()
}
\keyword{scale}
