% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plygnPrptns.R
\name{plygnPrptns}
\alias{plygnPrptns}
\title{Plot Density Plots}
\usage{
plygnPrptns(df, type, varR, varS, annual = F, col1 = "#F2AD00",
  col2 = "#67A9CF", col3 = "#F98400", col4 = "#FF0000", shade = "99")
}
\arguments{
\item{df}{Requires a dataframe with model simulation results. Must have a column of "var".}

\item{type}{Requires a character vector of "int" if plotting intercept data. Anything less (e.g., "exp") will default to plotting variance exponent data.}

\item{varR}{Observed numeric value of the variance exponent. Output generated from gather_density_data().}

\item{varS}{Simulated numeric values of the variance exponent. Output generated from gather_density_data().}

\item{annual}{Requires a T/F as to whether the data are annually summarized (T) or monthly summarized (F). Defaults to False.}

\item{col1}{Polygon color (character HEX) of simulated variance exponent parameters that are less extreme than the observed.}

\item{col3}{Polygon color (character HEX) of simulated variance exponent parameters that are more extreme than the observed.}

\item{col4}{Outline color  (character HEX) of simulated variance exponent parameters that are more extreme than the observed.}

\item{shade}{Requires a character object between 0 and 100. The lower the value the greater the transparency. Default is "99".}

\item{col1}{Line color (character HEX) of real variance exponent parameter.}
}
\description{
Plot proportional plots of the variance parameter depicting the simulated values greather than and less than the observed.
}
\examples{
plygnPrptns(df, type = "exp", varR = varR, varS = varS, col1 = "#F2AD00", col3 = "#F98400", col4 = "#FF0000")
}
\keyword{Density}
\keyword{Plot,}
