% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_SE.R
\name{pointsSE}
\alias{pointsSE}
\title{Add Points and Standard Errors}
\usage{
pointsSE(df, lcol = "#969696", cols, LW, shade = "99", ptsize = 1)
}
\arguments{
\item{df}{Requires a dataframe with slope (slope), standard error (se) and sqrt reponse variable (sqrt.resp) columns. The sqrt.resp variable column should be produced by the arrange.data() function.}

\item{lcol}{Requires a character object defining the color of the lines. The default is grey.}

\item{cols}{Requires a character vector of either 1 or the same length as the number of plot points. These values should be output from the color.gradient() function.}

\item{LW}{Requires a numeric value that defines the line width of the SE segments.}

\item{shade}{Requires a character object between 0 and 100. The lower the value the greater the transparency. Default is "99".}
}
\description{
This function allows you to add points and SE lines to plotting region.
}
\examples{
points.SE(df, lcol = "#969696", cols = cols, LW = 1, shade = "99")
}
\keyword{SE-lines}
\keyword{points,}
