% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{zero_one}
\alias{zero_one}
\title{0 to 1 scaling function.}
\usage{
zero_one(x)
}
\arguments{
\item{x}{Requires a vector of data.}
}
\description{
This function allows you to scale data between 0 and 1.
}
\examples{
zero_one()
}
\keyword{scale}
